Move a class to enum and add some comments to range calculation
